Data.Count.Counter

Documentation

data Counter a

A Counter a maps bijectively between a subset of values of type a and some possibly empty or infinite prefix of [0..].

cCount is Just n when the counter is finite and manages n values, or Nothing when infinite.

cToPos converts a managed value to its natural number (starting from 0).

cFromPos converts a natural number to its managed value.

cToPos c . cFromPos c must be the identity function. This invariant is maintained using the combinators below.

Constructors

UnsafeMkCounter

Fields

cCount :: Maybe Integer

cToPos :: a -> Integer

cFromPos :: Integer -> a

unitCounter :: Counter ()

A counter for the single unit value.

voidCounter :: Counter a

A counter for an empty set of values, for any type.

natCounter :: Counter Integer

Counts through the natural numbers: [0..] maps simply to [0..].

dropCounter :: Integer -> Counter a -> Counter a

dropCounter n c drops the first n elements from the given counter. cToPos (dropCounter n c) 0 is equivalent to cToPos c 0.

takeCounter :: Integer -> Counter a -> Counter a

If n is greater than the count of c, takeCounter n c creates a counter for the first n elements.

sumCounter :: Counter a -> Counter b -> Counter (Either a b)

Given two counters, a and b, creates a counter for all Left-tagged a values and Right-tagged b values.

prodCounter :: Counter a -> Counter b -> Counter (a, b)

Creates a counter for the Cartesian product of values in two given counters.

boundedEnumCounter :: (Bounded a, Enum a) => Counter a

A counter for any Bounded Enum. [minBound :: a ..] maps to [0..]. Invalid if fromEnum (minBound :: a) or fromEnum (maxBound :: a) overflows.

isoCounter :: Counter a -> (b -> a) -> (a -> b) -> Counter b

maybeCounter :: Counter a -> Counter (Maybe a)

listCounter :: Counter a -> Counter [a]

Counter for all lists of all values in given counter.

The count is 1 (the only value being the empty list) if the given counter is empty, infinite otherwise.

integerCounter :: Counter Integer

Maps [0,-1,1,-2,2,..] to [0..]

boundedIntegralCounter :: (Bounded a, Integral a) => Counter a

Similar to integerCounter, but bounded. Doesn't overflow, unlike boundedEnumCounter.

allValuesFor :: Counter a -> [a]

All values in the given counter, from the 0 correspondent upwards.